requirements.txt
langchain>=0.1.0
openai>=1.0.0
chromadb>=0.4.0
yfinance>=0.2.36
newsapi-python>=0.2.7
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=4.0.0
pypdf>=3.0.0
python-docx>=0.8.11
flask-uploads>=0.2.1
tiktoken>=0.5.1